requirements.txt
spotipy
python-dotenv
pandas
numpy
tensorflow
scikit-learn
matplotlib
keras
scipy
seaborn
mlxtend
pickle